Add dev container Dockerfiles and build workflow, and Dependabot
We're using AWS CodeBuild to do the building. GitHub actions was far too
slow for our case.

And in order to speed things up a bit - and in time, to share between
the code and dev builds - the parts of the build are split into separate
images which can be built in parallel with each other and skipped when
they haven't changed.

Add Dependabot to keep things up to date.