workflow: Add an overall status job
Branch protection rules are a bit silly and you can only wait on
individual jobs. That means (1) we have to enumerate them all (would be
ok with terraform probably), but (2) if any of them are skipped due to
our path filters then the PR would be blocked.

So we add a collection job at the end which reports the overall status,
and we can wait on that.