docker-build-and-push: Reference the public images if we skip building
If the PR doesn't change one of the Dockerfiles, it'll be skipped. In
that case the combined image must pull from the public repo to get the
latest changes. We weren't correctly doing that since we were just
looking for whether the trigger is a PR or not to decide whether to use
the temp or permanent repos.