Always log into both registries
I've set a secret in the codebuild project with the private registry
name, should stop any monkeying with figuring out which ones are needed.
Problem is when building the combined image you might need both: if one
got skipped because it didn't change in the PR then we're supposed to
fetch it from the public registry. But the workflow was only coded to
log in to one or the other.